requirements.txt
adal==1.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argcomplete==3.6.3
asgiref==3.9.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.38.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.6.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==25.1.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==22.0.0
azureml-core==1.54.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
contextlib2==21.6.0
contourpy==1.3.3
cryptography==41.0.7
cycler==0.12.1
Django==5.2.5
docker==6.1.3
fastapi==0.124.4
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
humanfriendly==10.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.1
jsonpickle==3.4.2
kagglehub==0.3.13
kiwisolver==1.4.8
knack==0.11.0
KNN==1.0.0
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.10.5
msal==1.34.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.6.4
ndg-httpsclient==0.5.1
numpy==2.3.2
oauthlib==3.3.1
packaging==24.0
pandas==2.3.1
paramiko==3.5.1
pathspec==0.12.1
pillow==11.3.0
pkginfo==1.12.1.2
playsound==1.2.2
plyer==2.1.0
portalocker==2.10.1
propcache==0.3.2
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
PyNaCl==1.6.2
pyOpenSSL==23.3.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
ruamel.yaml==0.17.40
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
SecretStorage==3.5.0
setuptools==81.0.0
six==1.17.0
sqlparse==0.5.3
starlette==0.50.0
tabulate==0.9.0
threadpoolctl==3.6.0
tk==0.1.0
tqdm==4.67.1
twilio==9.7.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.46.3
yarl==1.20.1
